Restructure interfaces and provide cross referencer
