+++ edu.kit.ipd.sdq.kamp.ruledsl/src/edu/kit/ipd/sdq/kamp/ruledsl/generator/KampRuleLanguageGenerator.xtend
@@ -261,11 +261,13 @@ class KampRuleLanguageGenerator implements IGenerator {
 			//requiredBundles.add("edu.kit.ipd.sdq.kamp4bp")
 			
 			// add the kamp packages here because we should resolve the bin folder dependencies via import not require
-			importedPackages.add("edu.kit.ipd.sdq.kamp4bp.ruledsl.support");
 			importedPackages.add("edu.kit.ipd.sdq.kamp.ruledsl.support");
+			importedPackages.add("edu.kit.ipd.sdq.kamp.propagation");
 			importedPackages.add("edu.kit.ipd.sdq.kamp4bp.core");	
 			importedPackages.add("edu.kit.ipd.sdq.kamp4is.core")
 			importedPackages.add("edu.kit.ipd.sdq.kamp4is.model.modificationmarks")
+			importedPackages.add("edu.kit.ipd.sdq.kamp.architecture")
+			importedPackages.add("edu.kit.ipd.sdq.kamp.model.modificationmarks");
 			
 			createManifest(projectName, requiredBundles, importedPackages, exportedPackages, monitor, project);
 	}
